+++ sql/core/src/test/scala/org/apache/spark/sql/jdbc/v2/JDBCV2JoinPushdownSuite.scala
@@ -22,7 +22,6 @@ import java.sql.Connection
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.{ExplainSuiteHelper, QueryTest}
 import org.apache.spark.sql.connector.DataSourcePushdownTestUtils
-import org.apache.spark.sql.execution.datasources.v2.jdbc.JDBCTableCatalog
 import org.apache.spark.sql.jdbc.{H2Dialect, JdbcDialect}
 import org.apache.spark.sql.test.SharedSparkSession
 import org.apache.spark.util.Utils
@@ -36,19 +35,12 @@ class JDBCV2JoinPushdownSuite
   val tempDir = Utils.createTempDir()
   override val url = s"jdbc:h2:${tempDir.getCanonicalPath};user=testUser;password=testPass"
 
-  override val catalogName: String = "h2"
   override val namespaceOpt: Option[String] = Some("test")
 
   override val jdbcDialect: JdbcDialect = H2Dialect()
 
   override def sparkConf: SparkConf = super.sparkConf
-    .set("spark.sql.catalog.h2", classOf[JDBCTableCatalog].getName)
-    .set("spark.sql.catalog.h2.url", url)
     .set("spark.sql.catalog.h2.driver", "org.h2.Driver")
-    .set("spark.sql.catalog.h2.pushDownAggregate", "true")
-    .set("spark.sql.catalog.h2.pushDownLimit", "true")
-    .set("spark.sql.catalog.h2.pushDownOffset", "true")
-    .set("spark.sql.catalog.h2.pushDownJoin", "true")
 
   override def qualifyTableName(tableName: String): String = namespaceOpt
     .map(namespace => s""""$namespace"."$tableName"""").getOrElse(s""""$tableName"""")
